requirements.txt
# WD Tagger Batch - Extension Requirements
# These are installed automatically by SD WebUI on first launch.

# ONNX inference (CPU)
onnxruntime>=1.16.0

# Image processing
Pillow>=9.0.0

# Model download from HuggingFace
huggingface-hub>=0.19.0

# Numpy (usually already installed)
numpy>=1.23.0
